requirements.txt
Flask>=2.0
psycopg2>=2.9
Werkzeug>=2.0
gunicorn>=20.0
python-dotenv>=0.19
flask-cors>=3.0
google-generativeai>=0.3.0